Added SVG export to PDF funtionality for health graphical
