+++ package.json
@@ -16,6 +16,7 @@
     "c3": "^0.7.20",
     "core-js": "^3.6.5",
     "d3": "^7.3.0",
+    "jspdf": "^2.5.1",
     "moment": "^2.29.1",
     "register-service-worker": "^1.7.1",
     "vue": "^2.6.11",
